# delivery-microservice/src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: YumYumNow Delivery Microservice API
  description: APIs for managing deliveries.
  version: "1.0"
servers:
  - url: https://api.server.test/delivery-microservice
...
paths:
...
  /restaurants/{restaurantId}/location:
    put:
      tags:
        - Restaurant
      summary: Update the location of Restaurant entity
      description: Admin and vendor can update the location of a given restaurant. This is an internal endpoint that ensures consistency. Should be called by other microservices when updating the location of a restaurant.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        description: Address
        content:
          application/json:
            schema:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
                format: double
              description: delivery address coordinates
        required: true
      responses:
        '200':
          description: Location update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          description: Bad Request. Invalid input format or missing parameters.
          content:
            application/json:
              schema:
                type: string
                description: Code indicating a bad request
                example: "BAD_REQUEST"
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                type: string
                description: Code indicating restaurant not found
                example: "RESTAURANT_NOT_FOUND"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: string
                description: Code indicating server error
                example: "SERVER_ERROR"
...
  /restaurants/{restaurantId}/new-orders:
    get:
      tags:
        - Restaurant
      summary: Gets orders for a restaurant that don't have a courier assigned.
      description: Returns a list of Delivery entities, that haven't been assigned or taken by any courier yet. This endpoint can be accessed only by a Vendor.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: List of not assigned Deliveries returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Delivery'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  parameters:
...
    UserId:
      name: userId
      description: User ID for authorization
      in: header
      required: true
      schema:
        type: string
...
    RestaurantId:
      name: restaurantId
      in: path
      description: ID of the Restaurant entity
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized access. User not found or not specified.
      content:
        application/json:
          schema:
            type: string
            description: Error message
    Forbidden:
      description: Forbidden. User lacks necessary permissions.
      content:
        application/json:
          schema:
            type: string
            description: Error message
...
  schemas:
    Delivery:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "deliveries")'
      x-field-extra-annotation: '@javax.persistence.OneToOne(mappedBy = "error", cascade = {javax.persistence.CascadeType.PERSIST, javax.persistence.CascadeType.MERGE, javax.persistence.CascadeType.REFRESH}) @javax.persistence.MapsId'
      properties:
        deliveryID:
          type: string
          x-field-extra-annotation: "@javax.persistence.Id"
          format: uuid
          description: Unique identifier for the Delivery (will be the same as orderID)
        courierID:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"courierId\")"
          type: string
          format: email
          description: Unique identifier for the Courier assigned to the Delivery (will be taken from the Courier entity)
        restaurantID:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"restaurantId\")"
          type: string
          format: email
          description: Unique identifier for the restaurant making the Delivery.
        customerID:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"customerId\")"
          type: string
          format: email
          description: Unique identifier for the customer ordering the Delivery.
        estimatedPrepTime:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"estimatedPrepTime\")"
          type: integer
          format: minutes
          description: Estimated time to cook/prepare the order
        error:
          $ref: '#/components/schemas/Error'
        status:
          $ref: "#/components/schemas/DeliveryStatus"
        pickup_time:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"pickupTime\")"
          type: string
          format: date-time
          description: Pick-up time in YYYY-MM-DD HH:mm format
          example: "2023-11-19 18:20"
        delivered_time:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"deliveredTime\")"
          type: string
          format: date-time
          description: Delivered time in YYYY-MM-DD HH:mm format
          example: "2023-11-19 18:20"
        deliveryAddress:
          x-field-extra-annotation: "@javax.persistence.ElementCollection @javax.persistence.Column(name = \"deliveryAddress\")"
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          description: Latitude and longitude of the location the delivery is going to
        currentLocation:
          x-field-extra-annotation: "@javax.persistence.ElementCollection @javax.persistence.Column(name = \"currentLocation\")"
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          description: Latitude and longitude of the current location of the delivery
        ratingCourier:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"ratingCourier\")"
          type: integer
          description: Rating of the delivery from 1 - 5 provided by the customer.
        ratingRestaurant:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"ratingRestaurant\")"
          type: integer
          description: Rating of the food from 1 - 5 provided by the customer.
    DeliveryStatus:
      type: string
      description: Updates on where the order is on its journey
      enum:
        - PENDING
        - ACCEPTED
        - REJECTED
        - PREPARING
        - GIVEN_TO_COURIER
        - ON_TRANSIT
        - DELIVERED
    Error:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name = \"errors\")"
      x-field-extra-annotation: '@javax.persistence.OneToOne @javax.persistence.JoinColumn(name = "error_id")'
      properties:
        errorId:
          type: string
          format: uuid
          x-field-extra-annotation: "@javax.persistence.Id"
          description: Unique identifier for the problem associated with the Delivery (will be the same as deliveryID)
        reason:
          type: string
          x-field-extra-annotation: "@javax.persistence.Column(name = \"reason\")"
          description: Reason (provided by courier) for problems with delivery
        value:
          type: integer
          x-field-extra-annotation: "@javax.persistence.Column(name = \"value\")"
          description: optional field to indicate delay in minutes
        type:
          $ref: "#/components/schemas/ErrorType"
        Delivery:
          $ref: "#/components/schemas/Delivery"
    ErrorType:
      type: string
      description: Type of problem with delivery (i.e., cancellation/delay). Can be scaled later. Potentially, delayed deliveries will also have a specified numeric delay.
      enum:
        - DELAYED
        - CANCELLED
        - NONE
        - OTHER
    Restaurant:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity"
      properties:
        restaurantID:
          type: string
          x-field-extra-annotation: "@javax.persistence.Id"
          format: email
          description: Unique identifier for the restaurant
        couriers:
          x-field-extra-annotation: "@javax.persistence.ElementCollection @javax.persistence.Column(name = \"couriers\")"
          type: array
          items:
            type: string
            description: List of couriers the restaurant prefers
        location:
          x-field-extra-annotation: "@javax.persistence.ElementCollection @javax.persistence.Column(name = \"location\")"
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          description: Longitude and latitude of the restaurant
        deliveryZone:
          x-field-extra-annotation: "@javax.persistence.Column(name = \"deliveryZone\")"
          type: number
          format: double
          description: Radius of the delivery zone
          default: 30
...